refactor(pipeline): extract shared utils from v2 lib to src/lib
Move time-utils.ts and calendar helper functions (buildHolidayExceptions,
computeHolidayEndDate, computeDateRange, calendarToServiceId) from
app-data-v2/lib/ to pipeline/src/lib/. This eliminates the v1→v2
reverse dependency where v1 ODPT builder imported from v2 lib.

- pipeline/src/lib/time-utils.ts: timeToMinutes, adjustOdptOvernightTimes
- pipeline/src/lib/odpt-calendar-utils.ts: shared ODPT calendar functions
- app-data-v2/lib/odpt/build-calendar.ts: retains only buildCalendarV2
